+++ ricaun.AppBundleTool/Program.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.IO;
 using System.IO.Compression;
+using System.Threading.Tasks;
 
 namespace ricaun.AppBundleTool
 {
@@ -54,21 +55,21 @@ namespace ricaun.AppBundleTool
                     {
                         appBundleName = appBundleInfoTemp.ApplicationPackage.Name;
 
-                        //Console.WriteLine($"Install: {appBundleInfoTemp.ApplicationPackage.AsString()}");
+                        var installMessage = $"Install: {appBundleInfoTemp.ApplicationPackage.AsString().ToConsoleGreen()}";
 
                         var appBundleInstalled = AppBundleUtils.FindAppBundleByAppName(appBundleName);
                         if (appBundleInstalled is not null)
                         {
                             applicationPluginsFolder = appBundleInstalled.AppBundleFolder.GetApplicationPlugins();
-                            //Console.WriteLine($"Replace: {appBundleInstalled.ApplicationPackage.AsString()}");
                         }
 
-                        appBundleInfoTemp.ToDataTable(appBundleInstalled).Print();
+                        // Print different between the appBundleInfoTemp and appBundleInstalled
+                        //appBundleInfoTemp.ToDataTable(appBundleInstalled)?.Print();
 
                         // Copy all files of the folder to a different folder.
-
                         if (appBundleInstalled is not null)
                         {
+                            Console.WriteLine($"Replace: {appBundleInstalled.ApplicationPackage.AsString().ToConsoleYellow()}");
                             DirectoryUtils.CopyFilesRecursively(appBundleInfoTemp.PathBundle, appBundleInstalled.PathBundle);
                         }
                         else
@@ -76,18 +77,7 @@ namespace ricaun.AppBundleTool
                             var appBundleFolder = Path.Combine(applicationPluginsFolder, appBundleInfoTemp.Name);
                             DirectoryUtils.CopyFilesRecursively(appBundleInfoTemp.PathBundle, appBundleFolder);
                         }
-
-                        //Console.WriteLine(bundlePathZip);
-
-                        //ApplicationPluginsUtils.DownloadBundle(applicationPluginsFolder, bundlePathZip, (ex) =>
-                        //{
-                        //    if (options.Verbosity)
-                        //        Console.WriteLine(ex);
-                        //}, (message) =>
-                        //{
-                        //    if (options.Verbosity)
-                        //        Console.WriteLine(message);
-                        //});
+                        Console.WriteLine(installMessage);
                     }
                     AppBundleUtils.FindAppBundleByAppName(appBundleName)?.ToDataTable().Print();
                 }
@@ -99,7 +89,7 @@ namespace ricaun.AppBundleTool
                         var appBundleInfoTemp = DownloadAppBundleInfo(bundleUrl);
                         if (appBundleInfoTemp is null)
                         {
-                            Console.WriteLine($"AppBundle '{bundleUrl}' not found.");
+                            Console.WriteLine($"AppBundle '{bundleUrl}' not found.".ToConsoleRed());
                             return;
                         }
 
@@ -109,20 +99,19 @@ namespace ricaun.AppBundleTool
                         }
                         else
                         {
-                            Console.WriteLine($"AppBundleInfo '{appBundleInfoTemp.Name}' not valid.");
+                            Console.WriteLine($"AppBundleInfo '{appBundleInfoTemp.Name}' not valid.".ToConsoleRed());
                             return;
                         }
                     }
 
-                    var appBundle = AppBundleUtils.FindAppBundleByAppName(appBundleName);
+                    var appBundle = AppBundleUtils.FindAppBundle(appBundleName);
                     if (appBundle is null)
                     {
-                        Console.WriteLine($"AppBundle '{appBundleName}' not found.");
+                        Console.WriteLine($"AppBundle '{appBundleName}' not found.".ToConsoleRed());
                         return;
                     }
                     var applicationPluginsFolder = appBundle.AppBundleFolder.GetApplicationPlugins();
-                    Console.WriteLine($"Uninstall: {appBundle.ApplicationPackage.AsString()}");
-                    //ApplicationPluginsUtils.DeleteBundle(applicationPluginsFolder, appBundle.Name);
+                    Console.WriteLine($"Uninstall: {appBundle.ApplicationPackage.AsString().ToConsoleRed()}");
 
                     UninstallAppBundle(appBundle);
                 }
@@ -132,26 +121,24 @@ namespace ricaun.AppBundleTool
                     var appBundleInfoTemp = DownloadAppBundleInfo(bundleUrl);
                     if (appBundleInfoTemp.IsValid())
                     {
-                        Console.WriteLine($"DownloadApp: {appBundleInfoTemp.ApplicationPackage.AsString()}");
                         appBundleName = appBundleInfoTemp.ApplicationPackage.Name;
                         var appBundleInfo = AppBundleUtils.FindAppBundleByAppName(appBundleName);
                         appBundleInfo?.Show();
                         if (appBundleInfo is null)
                         {
-                            Console.WriteLine($"AppBundle '{appBundleName}' not found.");
+                            Console.WriteLine($"AppBundle '{appBundleName}' not found.".ToConsoleRed());
                             return;
                         }
                     }
                     else
                     {
-                        Console.WriteLine($"AppBundleInfo '{appBundleInfoTemp.Name}' not valid.");
+                        Console.WriteLine($"AppBundleInfo '{appBundleInfoTemp.Name}' not valid.".ToConsoleRed());
                         return;
                     }
                 }
                 else
                 {
-                    var appBundle = AppBundleUtils.FindAppBundleByAppName(appBundleName) ?? AppBundleUtils.FindAppBundleByBundleName(appBundleName);
-
+                    var appBundle = AppBundleUtils.FindAppBundle(appBundleName);
                     if (appBundle is null)
                     {
                         Show();
@@ -199,16 +186,22 @@ namespace ricaun.AppBundleTool
         private static AppBundleInfo DownloadAppBundleInfo(string bundleUrl)
         {
             var bundleName = Path.GetFileName(bundleUrl);
-            var progressText = $"Download: {bundleName.ToConsoleGreen()}";
+            var downloadProgress = $"Download: {bundleName.ToConsoleGreen()}";
 
-            var bundlePathZip = DownloadUtils.DownloadAsync(bundleUrl).ConsoleWaitResult(progressText);
+            var bundlePathZip = DownloadUtils.DownloadAsync(bundleUrl).ConsoleWaitResult(downloadProgress);
 
             // unzip file to folder
             var bundlePathFolder = Path.Combine(Path.GetDirectoryName(bundlePathZip), Path.GetFileNameWithoutExtension(bundlePathZip));
             if (Directory.Exists(bundlePathFolder))
                 Directory.Delete(bundlePathFolder, true);
 
-            ZipFile.ExtractToDirectory(bundlePathZip, bundlePathFolder, true);
+            var extractProgress = $"Extract: {bundleName.ToConsoleGreen()}";
+            Task.Run(() =>
+            {
+                ZipFile.ExtractToDirectory(bundlePathZip, bundlePathFolder, true);
+                return true;
+            }).ConsoleWaitResult(extractProgress);
+
             return AppBundleInfo.FindAppBundle(bundlePathFolder);
         }